Some progress

+++ Assets/Assets/Scripts/Data/Node.cs
@@ -5,7 +5,7 @@ public class Node
 {
     public Vector2 Position;
     public int GlobalIndex;
-    public float Temperature;
+    public double Temperature;
 
 
     public Node(Vector2 position) : this(position, 20.0f)
@@ -13,13 +13,13 @@ public class Node
         Position = position;
     }
 
-    public Node(Vector2 position, float temperature)
+    public Node(Vector2 position, double temperature)
     {
         Position = position;
         Temperature = temperature;
     }
 
-    public Node(Vector2 position, int index, float temperature)
+    public Node(Vector2 position, int index, double temperature)
     {
         Position = position;
         GlobalIndex = index;
